# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema Bancário API",
        "description": "API completa para sistema bancário com operações de contas, transações, cartões, seguros e relatórios",
        "contact": {
            "name": "Suporte Técnico",
            "email": "admin@banco.ao"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/contas": {
            "get": {
                "tags": [
                    "Contas"
                ],
                "summary": "Listar contas",
                "operationId": "17f5f71e68d179d6fae0dcdc28d08e82",
                "parameters": [
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "agencia_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status_conta_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Criar conta",
                "operationId": "7a99dd180e0adf67a82eb4c4035a4abc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "agencia_id",
                                    "tipo_conta_id",
                                    "moeda_id",
                                    "status_conta_id"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer"
                                    },
                                    "agencia_id": {
                                        "type": "integer"
                                    },
                                    "tipo_conta_id": {
                                        "type": "integer"
                                    },
                                    "moeda_id": {
                                        "type": "integer"
                                    },
                                    "saldo": {
                                        "description": "Opcional. Normalmente 0",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status_conta_id": {
                                        "type": "integer"
                                    },
                                    "numero_conta": {
                                        "type": "string",
                                        "readOnly": true
                                    },
                                    "iban": {
                                        "type": "string",
                                        "readOnly": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta criada"
                    }
                }
            }
        },
        "/api/contas/{id}": {
            "get": {
                "tags": [
                    "Contas"
                ],
                "summary": "Obter conta",
                "operationId": "fb871d41824a6ecea8db26e29fe5cc1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contas"
                ],
                "summary": "Atualizar conta",
                "operationId": "0aaadb162fbf4f5bf1d89a8e65ecf16e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "agencia_id": {
                                        "type": "integer",
                                        "readOnly": true
                                    },
                                    "tipo_conta_id": {
                                        "type": "integer"
                                    },
                                    "moeda_id": {
                                        "type": "integer"
                                    },
                                    "status_conta_id": {
                                        "type": "integer"
                                    },
                                    "numero_conta": {
                                        "type": "string",
                                        "readOnly": true
                                    },
                                    "iban": {
                                        "type": "string",
                                        "readOnly": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conta atualizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contas"
                ],
                "summary": "Remover conta",
                "operationId": "6b33997d8dcf861a5bb49b5a5be729eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluída"
                    }
                }
            }
        },
        "/api/contas/{id}/depositar": {
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Depositar em conta",
                "operationId": "63baddbaca7a7e57aca32dc0dcec24b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor",
                                    "moeda_id"
                                ],
                                "properties": {
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "moeda_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Depósito em numerário"
                                    },
                                    "referencia_externa": {
                                        "type": "string",
                                        "example": "EXT-12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Depósito efetuado"
                                        },
                                        "transacao": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contas/{id}/levantar": {
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Levantamento em conta",
                "operationId": "cae765e31ca16ab2bd2dc32ec833fd96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor",
                                    "moeda_id"
                                ],
                                "properties": {
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "moeda_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Levantamento no balcão"
                                    },
                                    "referencia_externa": {
                                        "type": "string",
                                        "example": "EXT-98765"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Levantamento efetuado"
                                        },
                                        "transacao": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contas/{id}/pagar": {
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Efetuar pagamento a parceiro",
                "operationId": "7cf215d4e731b81ddd26ff26d887a90a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parceiro",
                                    "referencia",
                                    "valor",
                                    "moeda_id"
                                ],
                                "properties": {
                                    "parceiro": {
                                        "type": "string"
                                    },
                                    "referencia": {
                                        "type": "string"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "moeda_id": {
                                        "type": "integer"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pagamento efetuado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação Bearer obtido via endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}